Move more into traits, take ValueSpecification into the more upstream part of fuzzing
